Form4: ------ Changed the screen from PrimaryScreen to Screen.FromControl(Me) Changed Bounds to WorkingArea Moved first grid of labels and Button1 to 92
Form5:
------
Changed the screen from PrimaryScreen to Screen.FromControl(Me)
Changed Bounds to WorkingArea
Moved first grid of labels and Button1 to 92
Added Uplink Channel Number with TextBox4 as input for the same.
TextBox4 changes the WaterMark depending on the band selected.
Reduced the size of the Input Command Box and Increased the size of Output Command Box.

Form9:
------
Changed the screen from PrimaryScreen to Screen.FromControl(Me)
Changed Bounds to WorkingArea
Changed the size and location of Output Command Box

Signed-off-by: Arun Kumar <arun.kumar@smartantennatech.com>